+++ Assets/Scenes/Scripts/Models/goods/PopGood.cs
@@ -5,6 +5,6 @@ using UnityEngine;
 public class PopGood 
 {
     public int Good_id;
-    public float Stockpile;
-    public float MaxNeed;    
+    public int Stockpile;
+    public int MaxNeed;    
 }
